backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wikipicture-web"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "celery[redis]>=5.3",
    "redis>=5.0",
    "python-multipart>=0.0.9",
    "aiofiles>=23.0",
    "pydantic-settings>=2.2",
    "jinja2>=3.1",
    "Pillow>=10.0",
    "pillow-heif>=0.13",
    "wikipicture>=0.1.0",
    "python-magic>=0.4.27",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
